HaskBF.Parser

Description

Documentation

type Program = [Op]

A program is a list of Ops. One possible Op is a Loop of other Ops

data Op

Brainfuck operations

Constructors

IncP
Increment the byte pointer: ">"

DecP
Decrement the byte pointer: "<"

Inc
Increment the byte pointed by the pointer: "+"

Dec
Decrement the byte pointed by the pointer: "-"

PutByte
Write the byte pointed by the pointer (side-effect): "."

GetByte
Read a byte, write at the current location (side-effect): ","

Loop [Op]
Loop over Ops untill current == 0 : "[...]"

parseProgram :: ByteString -> Either ParseError Program

Parse program stream. Returns an error or the parsed Program